use custom Exception classes for specific failures; v0.2.0
Where possible we try to use an Exception class from the stdlib, such as `ArgumentError`. However, as we are throwing different errors frm the `SpaceCadet` package we want to make it easy for consumers to respond to failures.

This changes quite a few of the naked `raise` statements to include a custom class.