improve retry logic around LB config being immutable
When the Rackspace Cloud Load Balancers are updated, they enter an immutable state. This returns an error from the API, so we need to retry the request. Unfortunately, it seems that load balancers can be immutable for more than 5 seconds.

So let's add some actual retries instead of a single retry.